Rename instructions file
Unfortunately, GitHub doesn't do what you expect with a file named
README.html. It does correctly identify the file as the one which
should be display at the bottom of the list for its directory, but
instead of properly rendering the HTML it shows the raw markup. So
we avoid that ugliness by using a name which doesn't start with
"README."