Set per-user PYTHONPATH variable (OCECDR-5001)
Windows seems to be inconsistent about when it expands nested
environment variables recursively and when it doesn't, so we need
to set each user's PYTHONPATH environment variable explicitly,
without relying on having the APPDATA variable resolved on the
fly from a global environment variable containing "%APPDATA%" as
part of its value. Sometimes that worked, but not all the time.