Show "no versions" popup (OCECDR-5198)
The search dialog was hiding the popup alert which let the user
know that the selected document has no versions, making the user
think that XMetaL had hung. Changed the code to turn off the flag
for the search dialog to stay on top of the windows stack while
the popup alert is shown, and then restore that flag after the
popup alert is dismissed.