Make Spanish macro usable for English summaries (OCECDR-5235)
The users have started using the Spanish Link ID Swap macro for
documents which are not Spanish, inserting a TranslationOf element
temporarily to trick the software into thinking the document was
Spanish. They did this without letting us know, so when the macros
were rewritten to use the more efficient method of determining
whether a summary document was Spanish using the SummaryLanguage
element which it already had when building the context menu, this
trick was broken. Falling back on the less efficient method, using
the TranslationOf element.